backend/requirements.txt
# Minimal dependencies for cron job system
supabase>=1.0.0
python-dotenv>=0.19.0
cryptography>=3.4.0
pytz>=2021.1
httpx>=0.20.0


